backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg>=3.1.0
python-dotenv==1.0.0
openai>=1.3.0
Pillow>=10.2.0
reportlab>=4.0.7
weasyprint>=61.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1